Clean up old notebooks
