--- config/baseline.csv
@@ -1,77 +1,0 @@
-56,0.560240963855
-77,0.660606060606
-54,0.71686746988
-42,0.958083832335
-48,0.560240963855
-43,1.0
-60,0.578313253012
-61,0.982035928144
-62,0.384146341463
-63,0.80487804878
-64,0.406060606061
-49,1.0
-66,0.988023952096
-67,1.0
-68,0.994011976048
-69,0.994011976048
-52,0.620481927711
-53,0.946107784431
-24,0.94578313253
-25,1.0
-26,0.958083832335
-27,0.957831325301
-20,0.664670658683
-21,0.74251497006
-22,0.88622754491
-23,1.0
-46,0.976047904192
-47,0.193939393939
-44,0.988023952096
-45,0.676646706587
-28,0.976047904192
-29,1.0
-40,0.897590361446
-41,0.497005988024
-1,0.904191616766
-3,0.861445783133
-2,0.873493975904
-5,0.398773006135
-4,0.614457831325
-7,0.610778443114
-6,0.843373493976
-9,0.170731707317
-8,0.910179640719
-51,0.721212121212
-39,0.562874251497
-65,0.623456790123
-76,0.487951807229
-75,0.748502994012
-38,0.916167664671
-73,0.88622754491
-72,0.317365269461
-71,1.0
-70,0.674698795181
-59,0.467065868263
-58,0.790419161677
-11,0.403614457831
-10,0.5
-13,0.277108433735
-12,0.150943396226
-15,0.802395209581
-14,0.802395209581
-17,0.618181818182
-16,0.820359281437
-19,0.928143712575
-18,0.335365853659
-31,0.796407185629
-30,0.922155688623
-37,0.493975903614
-36,0.320987654321
-35,0.844311377246
-34,0.419161676647
-33,0.620481927711
-55,0.275449101796
-74,0.3
-32,0.940119760479
-57,0.491017964072
-50,0.590361445783
