--- config/location_bins.csv
@@ -1,165 +1,0 @@
-ABANDONED BUILDING,p
-AIRCRAFT,p
-AIRPORT BUILDING NON-TERMINAL - NON-SECURE AREA,
-AIRPORT BUILDING NON-TERMINAL - SECURE AREA,
-AIRPORT EXTERIOR - NON-SECURE AREA,
-AIRPORT EXTERIOR - SECURE AREA,
-AIRPORT PARKING LOT,
-AIRPORT TERMINAL LOWER LEVEL - NON-SECURE AREA,
-AIRPORT TERMINAL LOWER LEVEL - SECURE AREA,
-AIRPORT TERMINAL MEZZANINE - NON-SECURE AREA,
-AIRPORT TERMINAL UPPER LEVEL - NON-SECURE AREA,
-AIRPORT TERMINAL UPPER LEVEL - SECURE AREA,
-AIRPORT TRANSPORTATION SYSTEM (ATS),
-AIRPORT VENDING ESTABLISHMENT,
-AIRPORT/AIRCRAFT,
-ALLEY,
-ANIMAL HOSPITAL,
-APARTMENT,
-APPLIANCE STORE,
-ATHLETIC CLUB,
-ATM (AUTOMATIC TELLER MACHINE),
-AUTO,
-BANK,
-BANQUET HALL,
-BAR OR TAVERN,
-BARBER SHOP/BEAUTY SALON,
-BARBERSHOP,
-BASEMENT,
-BOAT/WATERCRAFT,
-BOWLING ALLEY,
-BRIDGE,
-CAR WASH,
-CEMETARY,
-CHA APARTMENT,
-CHA BREEZEWAY,
-CHA ELEVATOR,
-CHA GROUNDS,
-CHA HALLWAY,
-CHA HALLWAY/STAIRWELL/ELEVATOR,
-CHA LOBBY,
-CHA PARKING LOT,
-CHA PARKING LOT/GROUNDS,
-CHA PLAY LOT,
-CHA STAIRWELL,
-CHURCH,
-CHURCH PROPERTY,
-CHURCH/SYNAGOGUE/PLACE OF WORSHIP,
-CLEANING STORE,
-CLUB,
-COACH HOUSE,
-COIN OPERATED MACHINE,
-COLLEGE/UNIVERSITY GROUNDS,
-COLLEGE/UNIVERSITY RESIDENCE HALL,
-COMMERCIAL / BUSINESS OFFICE,
-CONSTRUCTION SITE,
-CONVENIENCE STORE,
-COUNTY JAIL,
-CREDIT UNION,
-CTA "L" PLATFORM,
-CTA "L" TRAIN,
-CTA BUS,
-CTA BUS STOP,
-CTA GARAGE / OTHER PROPERTY,
-CTA PLATFORM,
-CTA PROPERTY,
-CTA STATION,
-CTA TRACKS - RIGHT OF WAY,
-CTA TRAIN,
-CURRENCY EXCHANGE,
-DAY CARE CENTER,
-DELIVERY TRUCK,
-DEPARTMENT STORE,
-DRIVEWAY,
-DRIVEWAY - RESIDENTIAL,
-DRUG STORE,
-DUMPSTER,
-FACTORY,
-FACTORY/MANUFACTURING BUILDING,
-FEDERAL BUILDING,
-FIRE STATION,
-FOREST PRESERVE,
-FUNERAL PARLOR,
-GANGWAY,
-GARAGE,
-GARAGE/AUTO REPAIR,
-GAS STATION,
-GAS STATION DRIVE/PROP.,
-GOVERNMENT BUILDING/PROPERTY,
-GROCERY FOOD STORE,
-HALLWAY,
-HIGHWAY/EXPRESSWAY,
-HOSPITAL,
-HOSPITAL BUILDING/GROUNDS,
-HOTEL,
-HOTEL/MOTEL,
-HOUSE,
-JAIL / LOCK-UP FACILITY,
-JUNK YARD/GARBAGE DUMP,
-LAKE,
-LAKEFRONT/WATERFRONT/RIVERBANK,
-LAUNDRY ROOM,
-LIBRARY,
-LIQUOR STORE,
-LIVERY AUTO,
-LIVERY STAND OFFICE,
-LOADING DOCK,
-MEDICAL/DENTAL OFFICE,
-MOTEL,
-MOVIE HOUSE/THEATER,
-NEWSSTAND,
-NURSING HOME,
-NURSING HOME/RETIREMENT HOME,
-OFFICE,
-OTHER,
-OTHER COMMERCIAL TRANSPORTATION,
-OTHER RAILROAD PROP / TRAIN DEPOT,
-PARK PROPERTY,
-PARKING LOT,
-PARKING LOT/GARAGE(NON.RESID.),
-PAWN SHOP,
-POLICE FACILITY/VEH PARKING LOT,
-POOL ROOM,
-PORCH,
-PRAIRIE,
-PUBLIC GRAMMAR SCHOOL,
-PUBLIC HIGH SCHOOL,
-RAILROAD PROPERTY,
-RESIDENCE,
-RESIDENCE PORCH/HALLWAY,
-RESIDENCE-GARAGE,
-RESIDENTIAL YARD (FRONT/BACK),
-RESTAURANT,
-RETAIL STORE,
-RIVER,
-RIVER BANK,
-ROOMING HOUSE,
-SAVINGS AND LOAN,
-SCHOOL YARD,
-SCHOOL, PRIVATE, BUILDING,
-SCHOOL, PRIVATE, GROUNDS,
-SCHOOL, PUBLIC, BUILDING,
-SCHOOL, PUBLIC, GROUNDS,
-SEWER,
-SIDEWALK,
-SMALL RETAIL STORE,
-SPORTS ARENA/STADIUM,
-STAIRWELL,
-STREET,
-TAVERN,
-TAVERN/LIQUOR STORE,
-TAXI CAB,
-TAXICAB,
-TRAILER,
-TRUCK,
-TRUCKING TERMINAL,
-VACANT LOT,
-VACANT LOT/LAND,
-VEHICLE - DELIVERY TRUCK,
-VEHICLE NON-COMMERCIAL,
-VEHICLE-COMMERCIAL,
-VESTIBULE,
-WAREHOUSE,
-WOODED AREA,
-YARD,
-YMCA
